[stg] Dalamud 6 changes, still some minor bugs

+++ Orchestrion/IPlaybackController.cs
@@ -3,7 +3,7 @@ namespace Orchestrion
 {
     interface IPlaybackController
     {
-        int CurrentSong { get; }
+        ushort CurrentSong { get; }
         bool EnableFallbackPlayer { get; set; }
         void PlaySong(int songId);
         void StopSong();
